+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to lux will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.11] - 2026-03-23
+Added support for coloring N lines after a trigger match and loading profiles from local project directories.
+
+### Added
+- add next scope to color N lines after a trigger match
+
+### Other
+- local profiles
+
 ## [0.1.10] - 2026-03-20